Move sql2 bound write execution into exec write
